almost done navigator
