+++ index.ios.js
@@ -1,12 +1,7 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- * @flow
- */
+
 
 import React, { Component } from 'react';
-import gameCode from './loadNewGame.js';
-import WelcomeScreen from './welcomeScreen';
+import landingPage from './landingPage'
 
 import {
   AppRegistry,
@@ -18,53 +13,24 @@ import {
   TouchableHighlight
 } from 'react-native';
 
-
-export default class Cahoots extends Component {
+class Cahoots extends Component {
 
   render() {
     return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>
-         Cahoots!
-        </Text>
-        <Text style={styles.instructions}>
-          To get started, edit index.ios.js
-        </Text>
-        <Text style={styles.instructions}>
-          Press Cmd+R to reload,{'\n'}
-          Cmd+D or shake for dev menu
-        </Text>
-        <WelcomeScreen onButtonClick={this.onPressStartNewGame} text={"Start new game"}/>
-        <WelcomeScreen onButtonClick={this.onPressConnectGame} text={"Connect to an existing game"}/>
-      </View>
+      <Navigator
+        initialRoute=\{\{id: 'landingPage'}}
+        renderScene={this.navigatorRenderScene}/>
     );
   }
 
-  onPressStartNewGame(){
-    console.log("START NEW GAME PRESSEED");
+  navigatorRenderScene(route,navigator){
+    _navigator = navigator;
+    switch (route.id) {
+      case 'landingPage':
+        return (<landingPage navigator={navigator} title="Start new game or connect"/>)
+      }
+    }
   }
 
-  onPressConnectGame() {
-    console.log("CONNECT TO GAME PRESSED")
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
-  },
-});
 
 AppRegistry.registerComponent('Cahoots', () => Cahoots);
